small tweaks to code and documentation wording re: authorization
somewhat prompted by reading about authorization vs authentication, which are not the same thing (and yet are also very intertwined)

+++ R/auth.R
@@ -1,33 +1,36 @@
 # environment to store credentials
 .state <- new.env(parent = emptyenv())
 
-#' Authorize user using Oauth2.0 Credentials
+#' Authorize \code{googlesheets} to access user data from Google
 #'
-#' User will be directed to web browser and asked to sign into their Google
-#' account and grant googlesheets access permission to user data for Google
-#' Spreadsheets and Google Drive. User credentials will be cached in .httr-oauth
-#' in the current working directory.
+#' The user will be directed to a web browser, asked to sign in to their Google
+#' account, and to grant \code{googlesheets} access to user data for Google
+#' Spreadsheets and Google Drive. User credentials will be cached in a file
+#' named \code{.httr-oauth} in the current working directory.
 #'
-#' @param new_user set to \code{TRUE} if you want to wipe the slate clean and
-#'   re-authenticate with the same or different Google account
+#' Based on
+#' \href{https://github.com/hadley/httr/blob/master/demo/oauth2-google.r}{this
+#' demo} from \code{\link[httr]{httr}}.
 #'
-#' Based on \href{https://github.com/hadley/httr/blob/master/demo/oauth2-google.r}{this demo} from \code{httr}
+#' @param new_user logical, defaults to \code{FALSE}. Set to \code{TRUE} if you
+#'   want to wipe the slate clean and re-authenticate with the same or different
+#'   Google account.
 #'
 #' @export
 authorize <- function(new_user = FALSE) {
 
-  if(new_user & file.exists(".httr-oauth")) {
+  if(new_user && file.exists(".httr-oauth")) {
     message("Removing old credentials ...")
     file.remove(".httr-oauth")
   }
 
-  scope_list <- paste("https://spreadsheets.google.com/feeds",
-                      "https://docs.google.com/feeds")
+  scope_list <- c("https://spreadsheets.google.com/feeds",
+                  "https://docs.google.com/feeds")
 
-  client_id <- getOption("googlesheets.client_id")
-  client_secret <- getOption("googlesheets.client_secret")
-
-  googlesheets_app <- httr::oauth_app("google", client_id, client_secret)
+  googlesheets_app <-
+    httr::oauth_app("google",
+                    key = getOption("googlesheets.client_id"),
+                    secret = getOption("googlesheets.client_secret"))
 
   google_token <-
     httr::oauth2.0_token(httr::oauth_endpoints("google"), googlesheets_app,
@@ -35,8 +38,9 @@ authorize <- function(new_user = FALSE) {
 
   # check for validity so error is found before making requests
   # shouldn't happen if id and secret don't change
-  if("invalid_client" %in% unlist(google_token$credentials))
+  if("invalid_client" %in% unlist(google_token$credentials)) {
     message("Authorization error. Please check client_id and client_secret.")
+  }
 
   stopifnot(inherits(google_token, "Token2.0"))
 
@@ -56,6 +60,6 @@ get_google_token <- function() {
     authorize()
   }
 
-  token <- httr::config(token = .state$token)
+  httr::config(token = .state$token)
 
 }
